requirements.txt
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
stripe
gunicorn
django-heroku